Convert functional CPU into an object with threadsafe state access.
By converting to an object, it allows multiple CPUs to be created and allows
for a CPU to run in it's own goroutine. The state access functions are
guaranteed to run only at instruction boundaries, ensuring that state updated
or returned is atomic with respect to the virtual machine cyle.